requirements.txt
Django==2.2.5
dj-database-url==0.5.0
psycopg2-binary==2.8.3
djangorestframework==3.10.3
python-social-auth==0.3.6
social-auth-app-django==3.1.0
WeasyPrint==51
django-cors-headers==3.2.1

gunicorn==19.9.0
whitenoise==4.1.2
